New feature: Show author names for shared notes
The list of shared notes/comments was a mess, since it only showed "Note
one here", "Note two here", and so on...

It has now been enhanced to prefix each note with the name of the person
who wrote that note, such as Yourfriend: "I don't like that other guy."

This makes it very easy to see who wrote each comment, which helps you
decide whether you trust that opinion or not. Especially useful in large
guilds where lots of people use this addon!

PS: Note that the player name prefixed before each note is the name of
the person's "main character" as determined by DoIKnowYou. It will be
some character on their account, but may not be the exact one they wrote
the note from. That's because DoIKnowYou stores notes as if they were
written by each account's "main" character's name.