Fixed Lua error for guildless players
Whenever you left a guild, or logged into a guild-less player, you
received a Lua error caused by this addon.

An if-statement was simply incorrectly written by the original author.

Basically "if (not #count>0)" is interpreted by Lua as "if (false>0)",
due to Lua handling the "not"-statement as part of the "#count"
statement, which in turn causes an "attempt to compare number with
boolean" error.

That broken statement has been fixed by rewriting it as "if (#count<=0)",
since the intent of the original statement was to check if the table was
empty.