New feature: Show tooltip notes from other authors
Allows you to see "fallback notes" in the tooltip whenever you haven't
personally written anything about a player.

Go into the DoIKnowYou options, under the Tooltip tab, and specify a
comma-separated list of "trusted note authors". Now, when you hover over
other players in the world, it will first check if your account has
written a note about the player, as usual. If not, then it checks if
any of the trusted people have written a note about that player.

The list of trusted people MUST be people that you've seen in the list
of "shared notes" for various players (that's another new feature; the
ability to see WHO wrote each shared note). It's important to be aware
that the name displayed in that "shared notes" list isn't always the same
as the person you're interested in trusting. All notes are stored as the
"primary character" for their account, so if "Angryorc" notes are saved
as "Mainorc", then you'll have to specify "Mainorc" as the trusted player
to be able to see the notes written by "Angryorc" (and their "Mainorc").

Just look through the list of shared notes and you'll see the EXACT
names to specify as trusted if you want to see that person's notes.

Lastly, be aware that the tooltip ALWAYS prioritizes YOUR OWN NOTE, and
if none is found THEN it checks your list of trusted authors, in the
order you've specified them, and picks the FIRST found note (if any).

The tooltip will NEVER display ALL notes. It will ONLY display ONE note.
This is due to strict tooltip limitations by Blizzard. There's only enough
room to show one note! Be happy that this new, convenient feature exists.
And if you want more details (such as seeing ALL notes), then just target
that player and type "/diky" to view EVERYTHING about them, as usual!

Enjoy!